# chore(core): optimized VARCHAR equality function