--- a/core/src/main/java/io/questdb/cairo/VarcharTypeDriver.java
+++ b/core/src/main/java/io/questdb/cairo/VarcharTypeDriver.java
@@ -34,6 +34,7 @@ import static io.questdb.cairo.ColumnType.VARCHAR_AUX_SHL;
 
 public class VarcharTypeDriver implements ColumnTypeDriver {
     public static final VarcharTypeDriver INSTANCE = new VarcharTypeDriver();
+    public static final int VARCHAR_AUX_WIDTH_BYTES = 2 * Long.BYTES;
     // Maximum byte length that we can fully inline into auxiliary memory. In this case
     // there is no need to store any part of the string in data memory.
     // When the string is longer than this, we store the first few bytes in auxiliary memory,
@@ -41,7 +42,18 @@ public class VarcharTypeDriver implements ColumnTypeDriver {
     public static final int VARCHAR_MAX_BYTES_FULLY_INLINED = 9;
     // We store a prefix of this many bytes in auxiliary memory when the value is too large to inline.
     public static final int VARCHAR_INLINED_PREFIX_BYTES = 6;
+    public static final long VARCHAR_INLINED_PREFIX_MASK = (1L << 8 * VARCHAR_INLINED_PREFIX_BYTES) - 1L;
     public static final long VARCHAR_MAX_COLUMN_SIZE = 1L << 48;
+
+    private static final int HEADER_WIDTH_SHL = 4;
+    private static final int HEADER_FLAGS_WIDTH = 4;
+    private static final int HEADER_FLAGS_MASK = (1 << HEADER_FLAGS_WIDTH) - 1;
+    private static final int HEADER_FLAG_INLINED = 1;
+    private static final int HEADER_FLAG_ASCII = 2;
+    private static final int HEADER_FLAG_NULL = 4;
+    private static final int FULLY_INLINED_STRING_OFFSET = 1;
+    private static final int INLINED_PREFIX_OFFSET = 4;
+    private static final int INLINED_LENGTH_MASK = (1 << 4) - 1;
     // The exclusive limit on the byte length of a varchar value. The length is encoded in 28 bits.
     private static final int VARCHAR_LENGTH_LIMIT_BYTES = 1 << 28;
 
@@ -203,8 +215,17 @@ public class VarcharTypeDriver implements ColumnTypeDriver {
         return sequence.of(dataMemAddr + Integer.BYTES, dataMemAddr + Integer.BYTES + size(header), isAscii(header));
     }
 
+    /**
+     * Reads a UTF-8 value from a VARCHAR column.
+     *
+     * @param rowNum the row number to read
+     * @param dataMem base pointer of the data vector
+     * @param auxMem base pointer of the auxiliary vector
+     * @param ab whether to return the A or B flyweight
+     * @return a <code>Utf8Seqence</code> representing the value at <code>rowNum</code>
+     */
     public static Utf8Sequence getValue(long rowNum, MemoryR dataMem, MemoryR auxMem, int ab) {
-        final long auxOffset = rowNum << 4;
+        final long auxOffset = VARCHAR_AUX_WIDTH_BYTES * rowNum;
         int raw = auxMem.getInt(auxOffset);
         assert raw != 0;
         int flags = raw & 0x0f; // 4 bit flags
@@ -235,14 +256,24 @@ public class VarcharTypeDriver implements ColumnTypeDriver {
         return null;
     }
 
+    /**
+     * Reads a UTF-8 value from a VARCHAR column.
+     *
+     * @param auxAddr base pointer of the auxiliary vector
+     * @param dataAddr base pointer of the data vector
+     * @param rowNum the row number to read
+     * @param utf8view flyweight for the inlined string
+     * @param utf8SplitView flyweight for the split string
+     * @return utf8view or utf8SplitView loaded with the read value
+     */
     public static Utf8Sequence getValue(
             long auxAddr,
             long dataAddr,
-            long row,
-            DirectUtf8String utf8view,
+            long rowNum,
+            InlinedVarchar utf8view,
             Utf8SplitString utf8SplitView
     ) {
-        long auxEntry = auxAddr + (row << VARCHAR_AUX_SHL);
+        long auxEntry = auxAddr + VARCHAR_AUX_WIDTH_BYTES * rowNum;
         int raw = Unsafe.getUnsafe().getInt(auxEntry);
         assert raw != 0;
         int flags = raw & 0x0f; // 4 bit flags
@@ -256,10 +287,10 @@ public class VarcharTypeDriver implements ColumnTypeDriver {
 
         if ((flags & 1) == 1) {
             // inlined string
-            int size = (raw >> 4) & 0x0f;
-            return utf8view.of(auxEntry + 1, auxEntry + size + 1, ascii);
+            int size = (raw >> HEADER_FLAGS_WIDTH) & INLINED_LENGTH_MASK;
+            return utf8view.of(auxEntry + FULLY_INLINED_STRING_OFFSET, (byte) size, ascii);
         }
-        // string is split, prefix is in aux mem and the suffix is in data mem
+        // string is split, prefix is in aux mem and the full string is in data mem
         return utf8SplitView.of(
                 auxEntry + 4,
                 dataAddr + getDataOffset(auxEntry),
@@ -275,7 +306,7 @@ public class VarcharTypeDriver implements ColumnTypeDriver {
 
     @Override
     public long auxRowsToBytes(long rowCount) {
-        return rowCount << VARCHAR_AUX_SHL;
+        return VARCHAR_AUX_WIDTH_BYTES * rowCount;
     }
 
     @Override
@@ -307,8 +338,8 @@ public class VarcharTypeDriver implements ColumnTypeDriver {
                 ff,
                 fd,
                 fileName,
-                rowLo << VARCHAR_AUX_SHL,
-                rowHi << VARCHAR_AUX_SHL,
+                VARCHAR_AUX_WIDTH_BYTES * rowLo,
+                VARCHAR_AUX_WIDTH_BYTES * rowHi,
                 memoryTag,
                 opts
         );
@@ -328,11 +359,11 @@ public class VarcharTypeDriver implements ColumnTypeDriver {
     ) {
         long lo;
         if (rowLo > 0) {
-            lo = getDataOffset(auxMem, rowLo << VARCHAR_AUX_SHL);
+            lo = getDataOffset(auxMem, VARCHAR_AUX_WIDTH_BYTES * rowLo);
         } else {
             lo = 0;
         }
-        long hi = getDataVectorSize(auxMem, (rowHi - 1) << VARCHAR_AUX_SHL);
+        long hi = getDataVectorSize(auxMem, VARCHAR_AUX_WIDTH_BYTES * (rowHi - HEADER_FLAG_INLINED));
         dataMem.ofOffset(
                 ff,
                 dataFd,
@@ -346,12 +377,12 @@ public class VarcharTypeDriver implements ColumnTypeDriver {
 
     @Override
     public long getAuxVectorOffset(long row) {
-        return row << VARCHAR_AUX_SHL;
+        return VARCHAR_AUX_WIDTH_BYTES * row;
     }
 
     @Override
     public long getAuxVectorSize(long storageRowCount) {
-        return storageRowCount << VARCHAR_AUX_SHL;
+        return VARCHAR_AUX_WIDTH_BYTES * storageRowCount;
     }
 
     @Override
@@ -360,7 +391,7 @@ public class VarcharTypeDriver implements ColumnTypeDriver {
     }
 
     public long getDataVectorOffset(long auxMemAddr, long row) {
-        return getDataOffset(auxMemAddr + (row << VARCHAR_AUX_SHL));
+        return getDataOffset(auxMemAddr + VARCHAR_AUX_WIDTH_BYTES * row);
     }
 
     @Override
@@ -376,12 +407,12 @@ public class VarcharTypeDriver implements ColumnTypeDriver {
 
     @Override
     public long getDataVectorSizeAt(long auxMemAddr, long row) {
-        return getDataVectorSize(auxMemAddr + (row << VARCHAR_AUX_SHL));
+        return getDataVectorSize(auxMemAddr + VARCHAR_AUX_WIDTH_BYTES * row);
     }
 
     @Override
     public long getDataVectorSizeAtFromFd(FilesFacade ff, int auxFd, long row) {
-        long auxFileOffset = row << VARCHAR_AUX_SHL;
+        long auxFileOffset = VARCHAR_AUX_WIDTH_BYTES * row;
         if (row < 0) {
             return 0;
         }
@@ -465,16 +496,16 @@ public class VarcharTypeDriver implements ColumnTypeDriver {
                 tgtAuxAddr
         );
         dstDataMem.jumpTo(offset);
-        dstAuxMem.jumpTo(sortedTimestampsRowCount << VARCHAR_AUX_SHL);
+        dstAuxMem.jumpTo(VARCHAR_AUX_WIDTH_BYTES * sortedTimestampsRowCount);
     }
 
     @Override
     public long setAppendAuxMemAppendPosition(MemoryMA auxMem, long rowCount) {
         if (rowCount > 0) {
-            auxMem.jumpTo((rowCount - 1) << VARCHAR_AUX_SHL);
+            auxMem.jumpTo(VARCHAR_AUX_WIDTH_BYTES * (rowCount - HEADER_FLAG_INLINED));
             final long dataMemOffset = getDataVectorSize(auxMem.getAppendAddress());
             // Jump to the end of file to correctly trim the file
-            auxMem.jumpTo(rowCount << VARCHAR_AUX_SHL);
+            auxMem.jumpTo(VARCHAR_AUX_WIDTH_BYTES * rowCount);
             return dataMemOffset;
         }
         auxMem.jumpTo(0);