--- a/core/src/main/java/io/questdb/cairo/VarcharTypeDriver.java
+++ b/core/src/main/java/io/questdb/cairo/VarcharTypeDriver.java
@@ -45,7 +45,6 @@ public class VarcharTypeDriver implements ColumnTypeDriver {
     public static final long VARCHAR_INLINED_PREFIX_MASK = (1L << 8 * VARCHAR_INLINED_PREFIX_BYTES) - 1L;
     public static final long VARCHAR_MAX_COLUMN_SIZE = 1L << 48;
 
-    private static final int HEADER_WIDTH_SHL = 4;
     private static final int HEADER_FLAGS_WIDTH = 4;
     private static final int HEADER_FLAGS_MASK = (1 << HEADER_FLAGS_WIDTH) - 1;
     private static final int HEADER_FLAG_INLINED = 1;
@@ -55,7 +54,8 @@ public class VarcharTypeDriver implements ColumnTypeDriver {
     private static final int INLINED_PREFIX_OFFSET = 4;
     private static final int INLINED_LENGTH_MASK = (1 << 4) - 1;
     // The exclusive limit on the byte length of a varchar value. The length is encoded in 28 bits.
-    private static final int VARCHAR_LENGTH_LIMIT_BYTES = 1 << 28;
+    private static final int LENGTH_LIMIT_BYTES = 1 << 28;
+    private static final int DATA_LENGTH_MASK = LENGTH_LIMIT_BYTES - 1;
 
     /**
      * Appends UTF8 varchar type to the data and aux vectors.
@@ -69,27 +69,26 @@ public class VarcharTypeDriver implements ColumnTypeDriver {
         if (value != null) {
             int size = value.size();
             if (size <= VARCHAR_MAX_BYTES_FULLY_INLINED) {
-                // we can inline up to 9 bytes, which is what we do here
-                int flags = 1; // flags are 4 bits, 1 = inlined
+                int flags = HEADER_FLAG_INLINED;
                 if (value.isAscii()) {
-                    flags |= 2; // ascii flag
+                    flags |= HEADER_FLAG_ASCII;
                 }
                 // size is known to be at most 4 bits
-                auxMem.putByte((byte) ((size << 4) | flags));
+                auxMem.putByte((byte) ((size << HEADER_FLAGS_WIDTH) | flags));
                 auxMem.putVarchar(value, 0, size);
                 auxMem.skip(VARCHAR_MAX_BYTES_FULLY_INLINED - size);
                 offset = dataMem.getAppendOffset();
             } else {
-                if (size >= VARCHAR_LENGTH_LIMIT_BYTES) {
+                if (size >= LENGTH_LIMIT_BYTES) {
                     throw CairoException.critical(0).put("varchar value is too long [size=")
-                            .put(size).put(", max=").put(VARCHAR_LENGTH_LIMIT_BYTES).put(']');
+                            .put(size).put(", max=").put(LENGTH_LIMIT_BYTES).put(']');
                 }
 
                 int flags = 0;  // not inlined
                 if (value.isAscii()) {
-                    flags |= 2; // ascii flag
+                    flags |= HEADER_FLAG_ASCII;
                 }
-                auxMem.putInt((size << 4) | flags);
+                auxMem.putInt((size << HEADER_FLAGS_WIDTH) | flags);
                 auxMem.putVarchar(value, 0, VARCHAR_INLINED_PREFIX_BYTES);
                 offset = dataMem.putVarchar(value, 0, size);
                 if (offset >= VARCHAR_MAX_COLUMN_SIZE) {
@@ -98,8 +97,7 @@ public class VarcharTypeDriver implements ColumnTypeDriver {
                 }
             }
         } else {
-            // 4 = NULL
-            auxMem.putInt(4);
+            auxMem.putInt(HEADER_FLAG_NULL);
             auxMem.skip(VARCHAR_INLINED_PREFIX_BYTES);
             offset = dataMem.getAppendOffset();
         }
@@ -159,21 +157,20 @@ public class VarcharTypeDriver implements ColumnTypeDriver {
         // 3. non-empty string -> the size is non-zero
         assert Unsafe.getUnsafe().getInt(auxEntry) != 0;
 
-        return Unsafe.getUnsafe().getLong(auxEntry + 8L) >>> 16;
+        return Unsafe.getUnsafe().getLong(auxEntry + Long.BYTES) >>> 16;
     }
 
     public static long getDataVectorSize(MemoryR auxMem, long offset) {
         final int raw = auxMem.getInt(offset);
         assert raw != 0;
-        final int flags = raw & 0x0f; // 4 bit flags
+        final int flags = raw & HEADER_FLAGS_MASK;
         final long dataOffset = getDataOffset(auxMem, offset);
 
-        if ((flags & 4) == 4 || (flags & 1) == 1) {
-            // null flag is set or fully inlined value
+        if ((flags & HEADER_FLAG_NULL) == HEADER_FLAG_NULL || (flags & HEADER_FLAG_INLINED) == HEADER_FLAG_INLINED) {
             return dataOffset;
         }
         // size of the string at this offset
-        final int size = (raw >> 4) & 0xffffff;
+        final int size = (raw >> HEADER_FLAGS_WIDTH) & DATA_LENGTH_MASK;
         return dataOffset + size;
     }
 
@@ -228,18 +225,17 @@ public class VarcharTypeDriver implements ColumnTypeDriver {
         final long auxOffset = VARCHAR_AUX_WIDTH_BYTES * rowNum;
         int raw = auxMem.getInt(auxOffset);
         assert raw != 0;
-        int flags = raw & 0x0f; // 4 bit flags
+        int flags = raw & HEADER_FLAGS_MASK;
 
-        if ((flags & 4) == 4) {
-            // null flag is set
+        if ((flags & HEADER_FLAG_NULL) == HEADER_FLAG_NULL) {
             return null;
         }
 
-        boolean ascii = (flags & 2) == 2;
+        boolean ascii = (flags & HEADER_FLAG_ASCII) == HEADER_FLAG_ASCII;
 
         if ((flags & 1) == 1) {
             // inlined string
-            int size = (raw >> 4) & 0x0f;
+            int size = (raw >> HEADER_FLAGS_WIDTH) & INLINED_LENGTH_MASK;
             return ab == 1 ? auxMem.getVarcharA(auxOffset + 1, size, ascii) : auxMem.getVarcharB(auxOffset + 1, size, ascii);
         }
         // string is split, prefix is in auxMem and the suffix is in data mem
@@ -247,9 +243,9 @@ public class VarcharTypeDriver implements ColumnTypeDriver {
 
         if (utf8SplitString != null) {
             return utf8SplitString.of(
-                    auxMem.addressOf(auxOffset + 4),
+                    auxMem.addressOf(auxOffset + INLINED_PREFIX_OFFSET),
                     dataMem.addressOf(getDataOffset(auxMem, auxOffset)),
-                    (raw >> 4) & 0xffffff,
+                    (raw >> HEADER_FLAGS_WIDTH) & DATA_LENGTH_MASK,
                     ascii
             );
         }
@@ -276,25 +272,24 @@ public class VarcharTypeDriver implements ColumnTypeDriver {
         long auxEntry = auxAddr + VARCHAR_AUX_WIDTH_BYTES * rowNum;
         int raw = Unsafe.getUnsafe().getInt(auxEntry);
         assert raw != 0;
-        int flags = raw & 0x0f; // 4 bit flags
+        int flags = raw & HEADER_FLAGS_MASK;
 
-        if ((flags & 4) == 4) {
-            // null flag is set
+        if ((flags & HEADER_FLAG_NULL) == HEADER_FLAG_NULL) {
             return null;
         }
 
-        boolean ascii = (flags & 2) == 2;
+        boolean ascii = (flags & HEADER_FLAG_ASCII) == HEADER_FLAG_ASCII;
 
-        if ((flags & 1) == 1) {
+        if ((flags & HEADER_FLAG_INLINED) == HEADER_FLAG_INLINED) {
             // inlined string
             int size = (raw >> HEADER_FLAGS_WIDTH) & INLINED_LENGTH_MASK;
             return utf8view.of(auxEntry + FULLY_INLINED_STRING_OFFSET, (byte) size, ascii);
         }
         // string is split, prefix is in aux mem and the full string is in data mem
         return utf8SplitView.of(
-                auxEntry + 4,
+                auxEntry + INLINED_PREFIX_OFFSET,
                 dataAddr + getDataOffset(auxEntry),
-                (raw >> 4) & 0xffffff,
+                (raw >> HEADER_FLAGS_WIDTH) & DATA_LENGTH_MASK,
                 ascii
         );
     }
@@ -417,18 +412,18 @@ public class VarcharTypeDriver implements ColumnTypeDriver {
             return 0;
         }
         final int raw = readInt(ff, auxFd, auxFileOffset);
-        final int flags = raw & 0x0f; // 4 bit flags
+        final int flags = raw & HEADER_FLAGS_MASK;
 
         final int offsetLo = readInt(ff, auxFd, auxFileOffset + 8L);
         final int offsetHi = readInt(ff, auxFd, auxFileOffset + 12L);
         final long dataOffset = Numbers.encodeLowHighInts(offsetLo, offsetHi) >>> 16;
 
-        if ((flags & 4) == 4 || (flags & 1) == 1) {
+        if ((flags & HEADER_FLAG_NULL) == HEADER_FLAG_NULL || (flags & HEADER_FLAG_INLINED) == HEADER_FLAG_INLINED) {
             // null flag is set or fully inlined value
             return dataOffset;
         }
         // size of the string at this offset
-        final int size = (raw >> 4) & 0xffffff;
+        final int size = (raw >> HEADER_FLAGS_WIDTH) & DATA_LENGTH_MASK;
         return dataOffset + size;
     }
 
@@ -567,16 +562,16 @@ public class VarcharTypeDriver implements ColumnTypeDriver {
 
     private static long getDataVectorSize(long auxEntry) {
         final int raw = Unsafe.getUnsafe().getInt(auxEntry);
-        final int flags = raw & 0x0f; // 4 bit flags
+        final int flags = raw & HEADER_FLAGS_MASK;
         final long dataOffset = getDataOffset(auxEntry);
 
-        if ((flags & 4) == 4 || (flags & 1) == 1) {
-            // null flag is set or fully inlined value
+        if ((flags & HEADER_FLAG_NULL) == HEADER_FLAG_NULL || (flags & HEADER_FLAG_INLINED) == HEADER_FLAG_INLINED) {
             return dataOffset;
         }
         // size of the string at this offset
-        final int size = (raw >> 4) & 0xffffff;
-        assert size > 6 : String.format("size %,d <= 6, dataOffset %,d", size, dataOffset);
+        final int size = (raw >> HEADER_FLAGS_WIDTH) & DATA_LENGTH_MASK;
+        assert size > VARCHAR_MAX_BYTES_FULLY_INLINED : String.format("size %,d <= %d, dataOffset %,d",
+                size, VARCHAR_MAX_BYTES_FULLY_INLINED, dataOffset);
 
         return dataOffset + size;
     }