--- a/core/src/main/java/io/questdb/cairo/sql/PageAddressCacheRecord.java
+++ b/core/src/main/java/io/questdb/cairo/sql/PageAddressCacheRecord.java
@@ -46,8 +46,8 @@ public class PageAddressCacheRecord implements Record, Closeable {
     private final ObjList<SymbolTable> symbolTableCache = new ObjList<>();
     private final Utf8SplitString utf8SplitViewA = new Utf8SplitString();
     private final Utf8SplitString utf8SplitViewB = new Utf8SplitString();
-    private final DirectUtf8String utf8viewA = new DirectUtf8String();
-    private final DirectUtf8String utf8viewB = new DirectUtf8String();
+    private final InlinedVarchar utf8viewA = new InlinedVarchar();
+    private final InlinedVarchar utf8viewB = new InlinedVarchar();
     private int frameIndex;
     private PageAddressCache pageAddressCache;
     private long rowIndex;
@@ -423,7 +423,7 @@ public class PageAddressCacheRecord implements Record, Closeable {
     }
 
     @Nullable
-    private Utf8Sequence getVarchar(int columnIndex, DirectUtf8String utf8view, Utf8SplitString utf8SplitView) {
+    private Utf8Sequence getVarchar(int columnIndex, InlinedVarchar utf8view, Utf8SplitString utf8SplitView) {
         final long dataPageAddress = pageAddressCache.getPageAddress(frameIndex, columnIndex);
         final long auxPageAddress = pageAddressCache.getIndexPageAddress(frameIndex, columnIndex);
         if (auxPageAddress == 0) {