--- a/core/src/main/java/io/questdb/griffin/engine/functions/eq/EqVarcharFunctionFactory.java
+++ b/core/src/main/java/io/questdb/griffin/engine/functions/eq/EqVarcharFunctionFactory.java
@@ -100,7 +100,10 @@ public class EqVarcharFunctionFactory implements FunctionFactory {
 
         @Override
         public boolean getBool(Record rec) {
-            return negated != Utf8s.equalsNc(constant, arg.getVarcharA(rec));
+            // argument order to equalsNc is important: the first argument can be either an inlined or
+            // a split varchar. That implementation should choose the data access pattern that is optimal
+            // for it, and the constant implementation can easily adapt to both patterns.
+            return negated != Utf8s.equalsNc(arg.getVarcharA(rec), constant);
         }
 
         @Override