… core/src/main/java/io/questdb/griffin/engine/functions/eq/EqVarcharStrFunctionFactory.java
… core/src/main/java/io/questdb/std/str/InlinedVarchar.java
… core/src/main/java/io/questdb/std/str/Utf8Sequence.java
… core/src/main/java/io/questdb/std/str/Utf8SplitString.java
… core/src/main/java/io/questdb/std/str/Utf8String.java
… core/src/main/java/io/questdb/std/str/Utf8s.java
… core/src/test/java/io/questdb/test/cairo/VarcharEqualsTest.java
… core/src/test/java/io/questdb/test/cairo/VarcharTypeDriverTest.java
… core/src/test/java/io/questdb/test/cutlass/pgwire/PGJobContextTest.java
… core/src/test/java/io/questdb/test/griffin/ExplainPlanTest.java
… core/src/test/java/io/questdb/test/griffin/PageFrameCursorTest.java