This PR introduces a SWAR implementation of equality between a constant and a varchar column, or between two varchar columns. The current implementation already has the behavior of not touching the data vector for the bytes available in the auxiliary vector, so the performance on short strings isn't much affected.

I haven't done proper profiling to see if there's something easy to optimize further. I know we could simplify the logic if we added some padding bytes to the end of each data vector, so that we can always safely load a `long` from any offset that is less than the vector size.

The implementation adds methods to `Utf8Sequence` that are specialized for this use case. Should be taken into consideration when deciding whether to merge this work.

Here's my benchmarking data. Summary: 

* this branch is somewhat faster (~10%) for short strings (tested with 3, 9, and 10 bytes)
* significantly faster (~250%) for longer strings (tested with 110 bytes)

The optimizations are relevant only for strings of exactly the same size; otherwise there's a fast length-check path that doesn't compare any string contents.

```plain
create table x as (select 
  rnd_varchar('123', '123456789', '1234567890') a, 
  rnd_varchar('123', '123456789', '1234567890') b 
  from long_sequence(100000000)
  );

x where a = b

master
315 ms

branch
280 ms

x where a = '123'

master
250 ms 

branch 
230 ms

x where a = '123456789'

master
275 ms

branch
230 ms

x where a = '1234567890'

master
215 ms

branch
170 ms


create table x as (select 
  rnd_varchar(
    'oaneuhdz,pldyh,4l58dh,4ld8h,rlichlr ocyphdirl ,.pydi ,lrp.di,lxr.gpdyil9.dp,ilkgd. p,lkxd ,.l9kdx,l.9p8yigdk,p9ylgd',
    'oaneuhdz,pldyh,4l58dh,4ld8h,rlichlr ocyphdirl ,.pydi ,lrp.di,lxr.gpdyil9.dp,ilkgd. p,lkxd ,.l9kdx,l.9p8yigdk,p9ylga',
    'oaneuhdz,pldyh,4l58dh,4ld8h,rlichlr ocyphdirl ,.pydi ,lrp.di,lxr.gpdyil9.dp,ilkgd. p,lkxd ,.l9kdx,l.9p8yigdk,p9ylgb'
    ) a
  from long_sequence(10000000)
  );

x where a = 'oaneuhdz,pldyh,4l58dh,4ld8h,rlichlr ocyphdirl ,.pydi ,lrp.di,lxr.gpdyil9.dp,ilkgd. p,lkxd ,.l9kdx,l.9p8yigdk,p9ylgd'

master
130 ms

branch
55 ms
```