@ideoma 2024-03-19
#### [PR Coverage check]

:heart_eyes: **pass** : 76 / 80 (95.00%)




#### file detail

|   |path|covered line|new line|coverage|
|----|----|----|----|----|
|:large_blue_circle:|io/questdb/std/str/Utf8SplitString.java|9|11|81.82%|
|:large_blue_circle:|io/questdb/std/str/Utf8Sequence.java|9|10|90.00%|
|:large_blue_circle:|io/questdb/std/str/InlinedVarchar.java|15|16|93.75%|
|:large_blue_circle:|io/questdb/std/str/Utf8s.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/cairo/sql/PageAddressCacheRecord.java|2|2|100.00%|
|:large_blue_circle:|io/questdb/cairo/VarcharTypeDriver.java|15|15|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/eq/EqVarcharFunctionFactory.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/std/str/Utf8String.java|24|24|100.00%|