@ideoma 2024-03-27
#### [PR Coverage check]

:heart_eyes: **pass** : 112 / 128 (87.50%)




#### file detail

|   |path|covered line|new line|coverage|
|----|----|----|----|----|
|:large_blue_circle:|io/questdb/std/str/DirectUtf8Sequence.java|0|1|00.00%|
|:large_blue_circle:|io/questdb/std/str/Utf8Sequence.java|4|11|36.36%|
|:large_blue_circle:|io/questdb/std/str/Utf8SplitString.java|10|12|83.33%|
|:large_blue_circle:|io/questdb/std/str/InlinedVarchar.java|15|17|88.24%|
|:large_blue_circle:|io/questdb/cairo/VarcharTypeDriver.java|42|45|93.33%|
|:large_blue_circle:|io/questdb/std/str/Utf8String.java|22|23|95.65%|
|:large_blue_circle:|io/questdb/std/str/Utf8s.java|1|1|100.00%|
|:large_blue_circle:|io/questdb/cairo/sql/PageAddressCacheRecord.java|2|2|100.00%|
|:large_blue_circle:|io/questdb/griffin/FunctionParser.java|13|13|100.00%|
|:large_blue_circle:|io/questdb/griffin/engine/functions/eq/EqVarcharFunctionFactory.java|3|3|100.00%|